Spirng Security
user our own username and password in palce of default password given by
spring security. but it is not correct way, bcz in real time application
there will be lack of user so cofiguring each user in
applicaion.properties file is not possible so in next we do correct by
creating different package and classes.